+++ lib/vagrant-service-manager/plugin.rb
@@ -1,5 +1,6 @@
-# Loads all services
+# Loads all services and actions
 Dir["#{File.dirname(__FILE__)}/services/*.rb"].each { |f| require_relative f }
+Dir["#{File.dirname(__FILE__)}/action/*.rb"].each { |f| require_relative f }
 
 module Vagrant
   module ServiceManager
@@ -18,7 +19,14 @@ module Vagrant
       end
 
       action_hook(:servicemanager, :machine_action_up) do |hook|
+        hook.before(VagrantPlugins::ProviderVirtualBox::Action::Network, setup_network)
         hook.append(Service::Docker)
+      end
+
+      def self.setup_network
+        Vagrant::Action::Builder.new.tap do |b|
+          b.use Action::SetupNetwork
+        end
       end
     end
   end